feat and fix: add task and fix auth

+++ backend/src/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request, Req, BadRequestException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -15,8 +15,8 @@ export class UserController {
   }
 
   @Get()
-  findAll() {
-    return this.userService.findAll();
+  findAll(@Request() req) {
+    return req.user
   }
 
   @Get(':id')
@@ -30,7 +30,12 @@ export class UserController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.userService.remove(id); // ⬅️ без +
+  remove(@Param('id') id: string, @Req() req) {
+    const currentUserId = req.user.userId; // или req.user.id, в зависимости от JwtStrategy
+
+    if (id === currentUserId) {
+      throw new BadRequestException('You cannot delete yourself');
+    }
+    return this.userService.remove(id);
   }
 }
